Refactor server-api module
