+++ server-api/build.gradle
@@ -8,24 +8,23 @@ ext {
     lombokMapstructBindingVersion = "0.2.0"
 }
 
-
 dependencies {
 
     implementation project(":server-domain")
-    implementation project(":server-data")
     implementation project(":server-infrastructure")
 
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springframework.boot:spring-boot-starter-validation")
-    implementation("org.springframework.boot:spring-boot-starter-test")
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
-    implementation("com.querydsl:querydsl-jpa")
+
     implementation("commons-io:commons-io:2.11.0")
-    testImplementation("io.rest-assured:rest-assured")
-    testImplementation("com.github.tomakehurst:wiremock-jre8:2.35.1")
     implementation("org.apache.httpcomponents.client5:httpclient5:5.2.1")
 
     implementation "org.mapstruct:mapstruct:${mapstructVersion}"
     annotationProcessor "org.mapstruct:mapstruct-processor:${mapstructVersion}"
     annotationProcessor "org.projectlombok:lombok-mapstruct-binding:${lombokMapstructBindingVersion}"
+
+    testImplementation("org.springframework.boot:spring-boot-starter-test")
+    testImplementation("io.rest-assured:rest-assured")
+    testImplementation("com.github.tomakehurst:wiremock-jre8:2.35.1")
 }
