+++ server-api/src/main/java/io/joyoungc/api/configuration/ServerApiWebConfig.java
@@ -1,5 +1,6 @@
 package io.joyoungc.api.configuration;
 
+
 import io.joyoungc.api.configuration.format.TimestampFormatAnnotationFormatterFactory;
 import io.joyoungc.api.configuration.messages.JsonMessageSource;
 import org.springframework.context.MessageSource;
@@ -14,6 +15,7 @@ public class ServerApiWebConfig implements WebMvcConfigurer {
     public void addFormatters(FormatterRegistry registry) {
         TimestampFormatAnnotationFormatterFactory factory = new TimestampFormatAnnotationFormatterFactory();
         registry.addFormatterForFieldAnnotation(factory);
+
     }
 
     @Bean
