+++ server-api/src/main/java/io/joyoungc/api/configuration/format/LocalDateTimeFormatter.java
@@ -5,7 +5,6 @@ import org.springframework.format.Formatter;
 import java.text.ParseException;
 import java.time.Instant;
 import java.time.LocalDateTime;
-import java.util.Date;
 import java.util.Locale;
 import java.util.TimeZone;
