+++ server-api/src/main/java/io/joyoungc/api/infrastructure/repository/MemberJpaRepository.java
@@ -1,0 +1,10 @@
+package io.joyoungc.api.infrastructure.repository;
+
+import io.joyoungc.data.jpa.domain.MemberEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+/***
+ * Created by Aiden Jeong on 2021.12.13
+ */
+public interface MemberJpaRepository extends JpaRepository<MemberEntity, Long>, MemberQueryDslRepository {
+}
