+++ server-api/src/main/java/io/joyoungc/api/infrastructure/repository/MemberQueryDslRepository.java
@@ -1,0 +1,13 @@
+package io.joyoungc.api.infrastructure.repository;
+
+import io.joyoungc.data.jpa.domain.MemberEntity;
+import io.joyoungc.domain.member.Grade;
+
+import java.util.List;
+
+/***
+ * Created by Aiden Jeong on 2022.02.24
+ */
+public interface MemberQueryDslRepository {
+    List<MemberEntity> findMembers(Grade grade);
+}
