+++ server-api/src/main/java/io/joyoungc/api/request/CreateOrderRequest.java
@@ -1,0 +1,18 @@
+package io.joyoungc.api.request;
+
+
+import io.joyoungc.domain.order.OrderStatus;
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.validation.constraints.NotNull;
+
+@Getter
+@Setter
+public class CreateOrderRequest {
+    @NotNull
+    private Long memberId;
+    @NotNull
+    private Long productId;
+    private OrderStatus status;
+}
