+++ server-api/src/main/java/io/joyoungc/api/response/OrderResponse.java
@@ -1,0 +1,16 @@
+package io.joyoungc.api.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class OrderResponse {
+    private Long orderId;
+    private Long discountPrice;
+    private String productName;
+}
