+++ server-api/src/main/resources/config/appication-default.yml
@@ -1,12 +1,3 @@
-spring:
-  datasource:
-    url: jdbc:h2:mem:testdb;DB_CLOSE_ON_EXIT=FALSE
-    driver-class-name: org.h2.Driver
-    username: api
-    password: pass
-  jpa:
-    database-platform: org.hibernate.dialect.H2Dialect
-
 server-api:
   redis:
     enabled: false
