+++ server-api/src/test/java/io/joyoungc/api/ServerApiApplicationTests.java
@@ -23,7 +23,6 @@ class ServerApiApplicationTests {
     @Test
     void test_properties() {
         Assertions.assertThat(env.getProperty("server.port")).isEqualTo("9010");
-        Assertions.assertThat(env.getProperty("spring.output.ansi.enabled")).isEqualTo("always");
         Assertions.assertThat(env.getProperty("logging.level.jdbc.sqlonly")).isEqualTo("info");
     }
