+++ server-api/src/test/java/io/joyoungc/api/application/rest/OrderIntegrationTest.java
@@ -1,0 +1,99 @@
+package io.joyoungc.api.application.rest;
+
+import io.joyoungc.api.TestJpaConfig;
+import io.joyoungc.api.request.CreateOrderRequest;
+import io.joyoungc.api.response.OrderResponse;
+import io.joyoungc.domain.member.Grade;
+import io.joyoungc.domain.member.Member;
+import io.joyoungc.domain.member.MemberRepository;
+import io.joyoungc.domain.order.OrderStatus;
+import io.joyoungc.domain.product.Product;
+import io.joyoungc.domain.product.ProductRepository;
+import io.restassured.http.ContentType;
+import io.restassured.response.Response;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.junit.jupiter.api.Order;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+
+import java.util.List;
+
+import static io.joyoungc.api.TestUtils.createUrlWithPort;
+import static io.restassured.RestAssured.given;
+import static org.assertj.core.api.Assertions.assertThat;
+
+/***
+ * Created by Aiden Jeong on 2022.04.03
+ */
+@Slf4j
+@TestJpaConfig
+@RequiredArgsConstructor
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+class OrderIntegrationTest {
+
+    @LocalServerPort
+    private int port;
+    private final MemberRepository memberRepository;
+    private final ProductRepository productRepository;
+
+    private static final String API_ENDPOINT = "/api/orders";
+
+    static Long memberId;
+    static Long productId;
+
+    // @BeforeEach
+    private void init() {
+        Member member = new Member();
+        member.setName("이름");
+        member.setGrade(Grade.VIP);
+        memberId = memberRepository.save(member);
+        productId = productRepository.save(new Product("상품", 10000L));
+    }
+
+    @Test
+    @Order(1)
+    void create_order() {
+        init();
+        String url = createUrlWithPort(API_ENDPOINT, port);
+        CreateOrderRequest create = new CreateOrderRequest();
+        create.setMemberId(memberId);
+        create.setProductId(productId);
+        create.setStatus(OrderStatus.ORDER);
+
+        given().log().all()
+                .contentType(ContentType.JSON)
+                .body(create)
+                .when().post(url)
+                .then().log().body()
+                .statusCode(200);
+    }
+
+    @Test
+    @Order(2)
+    void get_member_orders() {
+        String url = createUrlWithPort(API_ENDPOINT, port) + "?memberId=" + memberId;
+
+        // given
+        Response response = given().log().all()
+                // when
+                .when().get(url)
+                // then
+                .then()
+                .log().body()
+                .statusCode(200)
+                .extract().response();
+
+        List<OrderResponse> list = response.jsonPath().getList(".", OrderResponse.class);
+
+        assertThat(list).isNotEmpty().element(0).satisfies(
+                c -> {
+                    assertThat(c.getOrderId()).isNotNull();
+                    assertThat(c.getProductName()).isEqualTo("상품");
+                    assertThat(c.getDiscountPrice()).isNotNull();
+                }
+        );
+    }
+
+}
